# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ShopEasy Gateway API
  version: 1.0.0
  description: |
    OpenAPI description for the gateway-facing ShopEasy API.

    Notes:
    - Browser clients must fetch `GET /api/csrf-token` and send `X-CSRF-Token` on mutating requests.
    - Authenticated routes use `Authorization: Bearer <access_token>`.
    - Guest cart routes use `X-Cart-ID` when a guest cart has already been created.
servers:
  - url: http://localhost
    description: Local storefront + gateway via Nginx
  - url: http://localhost:5000
    description: Direct gateway access
...
paths:
...
  /api/register:
    post:
      tags: [Auth]
      summary: Register a new user
      security:
        - csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccessResponse'
        '400':
          description: Validation error or duplicate email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '403':
          description: Missing or invalid CSRF token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitResponse'
  /api/login:
    post:
      tags: [Auth]
      summary: Login and receive JWT tokens
      security:
        - csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccessResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '403':
          description: Missing or invalid CSRF token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitResponse'
  /api/logout:
    post:
      tags: [Auth]
      summary: Logout and blacklist the access token
      security:
        - bearerAuth: []
        - csrfToken: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                required: [success]
        '401':
          description: Missing or invalid authorization header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '403':
          description: Missing or invalid CSRF token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/profile:
    put:
      tags: [Auth]
      summary: Update the authenticated user's profile
      security:
        - bearerAuth: []
        - csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileUpdateRequest'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required: [success]
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
  /api/products:
    get:
      tags: [Products]
      summary: List products with filtering and pagination
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
        - in: query
          name: page_size
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: min_price
          schema:
            type: number
        - in: query
          name: max_price
          schema:
            type: number
        - in: query
          name: sort_by
          schema:
            type: string
            enum: [id, name, price, category, stock, created_at]
        - in: query
          name: sort_order
          schema:
            type: string
            enum: [asc, desc]
      responses:
        '200':
          description: Product page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
  /api/products/{product_id}:
    get:
      tags: [Products]
      summary: Get a single product
      parameters:
        - in: path
          name: product_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/cart:
    get:
      tags: [Cart]
      summary: Get the current guest or authenticated cart
      parameters:
        - $ref: '#/components/parameters/XCartId'
      responses:
        '200':
          description: Cart payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
  /api/cart/add/{product_id}:
    post:
      tags: [Cart]
      summary: Add a product to the cart
      security:
        - csrfToken: []
      parameters:
        - in: path
          name: product_id
          required: true
          schema:
            type: integer
        - $ref: '#/components/parameters/XCartId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                quantity:
                  type: integer
                  minimum: 1
      responses:
        '200':
          description: Item added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartMutationResponse'
        '403':
          description: Missing or invalid CSRF token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/cart/remove/{product_id}:
    post:
      tags: [Cart]
      summary: Remove quantity or remove an item from the cart
      security:
        - csrfToken: []
      parameters:
        - in: path
          name: product_id
          required: true
          schema:
            type: integer
        - $ref: '#/components/parameters/XCartId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                quantity:
                  type: integer
                  minimum: 1
      responses:
        '200':
          description: Item removed or reduced
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  cart_count:
                    type: integer
                required: [success, cart_count]
  /api/cart/clear:
    post:
      tags: [Cart]
      summary: Clear the current cart
      security:
        - csrfToken: []
      parameters:
        - $ref: '#/components/parameters/XCartId'
      responses:
        '200':
          description: Cart cleared
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required: [success]
  /api/cart/count:
    get:
      tags: [Cart]
      summary: Get total item count in the cart
      parameters:
        - $ref: '#/components/parameters/XCartId'
      responses:
        '200':
          description: Cart count
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                required: [count]
  /api/cart/merge:
    post:
      tags: [Cart]
      summary: Merge a guest cart into the authenticated user's cart
      security:
        - bearerAuth: []
        - csrfToken: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                guest_cart_id:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Merge result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  cart_id:
                    type: string
                    format: uuid
                  message:
                    type: string
                required: [success]
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
...
  schemas:
...
    AuthSuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        user:
          $ref: '#/components/schemas/UserSummary'
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
        expires_in:
          type: integer
      required: [success, user, access_token, refresh_token, token_type, expires_in]
    AuthErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        error:
          type: string
      required: [success, error]
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
      required: [name, email, password]
...