# services/gateway/app.py
"""
API Gateway for Microservices

Routes requests to backend services with:
- JWT token validation
- Rate limiting (Redis-backed)
- Request routing
- CORS handling
- Security headers
"""
import os
import time
import uuid
from datetime import datetime, timedelta
from flask import Flask, request, jsonify, Response, g
from flask_cors import CORS
import redis
import jwt
from functools import wraps
import requests

app = Flask(__name__)

# Configuration
SECRET_KEY = os.environ.get('JWT_SECRET_KEY') or os.environ.get('SECRET_KEY')
if not SECRET_KEY:
    raise RuntimeError('JWT_SECRET_KEY or SECRET_KEY environment variable is required')
REDIS_URL = os.environ.get('REDIS_URL', 'redis://redis:6379/0')

# Service URLs
SERVICES = {
    'auth': os.environ.get('AUTH_SERVICE_URL', 'http://auth-service:5001'),
    'product': os.environ.get('PRODUCT_SERVICE_URL', 'http://product-service:5002'),
    'cart': os.environ.get('CART_SERVICE_URL', 'http://cart-service:5003'),
    'order': os.environ.get('ORDER_SERVICE_URL', 'http://order-service:5004'),
    'inventory': os.environ.get('INVENTORY_SERVICE_URL', 'http://inventory-service:5005'),
    'legacy': os.environ.get('LEGACY_SERVICE_URL', 'http://application:5000'),
}

# Routes that don't require authentication
PUBLIC_ROUTES = [
    ('POST', '/api/auth/login'),
    ('POST', '/api/auth/register'),
    ('POST', '/api/auth/refresh'),
    ('GET', '/api/products'),
    ('GET', '/api/products/'),  # Product details
    ('GET', '/api/categories'),
    ('GET', '/health'),
    ('GET', '/ready'),
    ('GET', '/live'),
]

# Routes that require authentication
PROTECTED_ROUTES = [
    ('POST', '/api/auth/logout'),
    ('GET', '/api/auth/me'),
    ('PUT', '/api/auth/profile'),
    ('GET', '/api/cart'),
    ('POST', '/api/cart/add'),
    ('POST', '/api/cart/remove'),
    ('POST', '/api/cart/clear'),
    ('GET', '/api/cart/count'),
    ('GET', '/api/orders'),
    ('POST', '/api/orders'),
    ('GET', '/api/orders/'),
]

# Enable CORS
CORS(app, supports_credentials=True)

# Redis client for rate limiting
redis_client = None

RATE_LIMIT_RULES = {
    ('POST', '/api/login'): {'limit': 10, 'window': 60},
    ('POST', '/api/register'): {'limit': 5, 'window': 60},
    ('POST', '/api/auth/login'): {'limit': 10, 'window': 60},
    ('POST', '/api/auth/register'): {'limit': 5, 'window': 60},
    ('POST', '/api/auth/refresh'): {'limit': 30, 'window': 60},
}


def get_redis():
    """Get Redis client"""
    global redis_client
    if redis_client is None:
        try:
            redis_client = redis.from_url(REDIS_URL)
        except Exception:
            pass
    return redis_client


def check_rate_limit(ip: str, endpoint: str, limit: int = 100, window: int = 60) -> tuple:
    """
    Check rate limit for IP + endpoint.
    Returns (allowed: bool, remaining: int, reset_time: int)
    """
    r = get_redis()
    if not r:
        return True, limit, int(time.time()) + window

    key = f"rate_limit:{ip}:{endpoint}"
    try:
        current = r.get(key)
        if current is None:
            r.setex(key, window, 1)
            return True, limit - 1, int(time.time()) + window

        current = int(current)
        if current >= limit:
            ttl = r.ttl(key)
            return False, 0, int(time.time()) + max(ttl, 0)

        r.incr(key)
        return True, limit - current - 1, int(time.time()) + r.ttl(key)
    except Exception:
        return True, limit, int(time.time()) + window


def get_rate_limit_rule(method: str, path: str):
    """Return the configured rate-limit rule for a request, if any."""
    return RATE_LIMIT_RULES.get((method, path))


def verify_jwt_token(token: str) -> dict:
    """Verify JWT token and return payload"""
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=['HS256'])

        # Check if token is blacklisted
        r = get_redis()
        if r and r.exists(f"blacklist:{token}"):
            return None

        return payload
    except jwt.ExpiredSignatureError:
        return None
    except jwt.InvalidTokenError:
        return None


def create_csrf_token() -> str:
    """Create a short-lived CSRF token for browser clients."""
    payload = {
        'type': 'csrf',
        'iat': datetime.utcnow(),
        'exp': datetime.utcnow() + timedelta(hours=12),
    }
    return jwt.encode(payload, SECRET_KEY, algorithm='HS256')


def verify_csrf_token(token: str) -> bool:
    """Validate a CSRF token generated by the gateway."""
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=['HS256'])
        return payload.get('type') == 'csrf'
    except jwt.InvalidTokenError:
        return False


def require_auth(f):
    """Decorator to require authentication"""
    @wraps(f)
    def decorated_function(*args, **kwargs):
        # Get token from Authorization header
        auth_header = request.headers.get('Authorization', '')
        if not auth_header.startswith('Bearer '):
            return jsonify({'error': 'Missing or invalid Authorization header'}), 401

        token = auth_header[7:]  # Remove 'Bearer ' prefix
        payload = verify_jwt_token(token)

        if not payload:
            return jsonify({'error': 'Invalid or expired token'}), 401

        # Store user info in request context
        g.user_id = payload.get('user_id')
        g.user_email = payload.get('email')
        g.user_name = payload.get('name')
        g.user_is_admin = bool(payload.get('is_admin'))

        return f(*args, **kwargs)
    return decorated_function


def require_admin(f):
    """Decorator to require an admin JWT token."""
    @wraps(f)
    def decorated_function(*args, **kwargs):
        auth_header = request.headers.get('Authorization', '')
        if not auth_header.startswith('Bearer '):
            return jsonify({'error': 'Missing or invalid Authorization header'}), 401

        token = auth_header[7:]
        payload = verify_jwt_token(token)
        if not payload:
            return jsonify({'error': 'Invalid or expired token'}), 401
        if not payload.get('is_admin'):
            return jsonify({'error': 'Admin access required'}), 403

        g.user_id = payload.get('user_id')
        g.user_email = payload.get('email')
        g.user_name = payload.get('name')
        g.user_is_admin = True
        return f(*args, **kwargs)
    return decorated_function


@app.before_request
def before_request():
    """Pre-request processing: logging, rate limiting"""
    g.request_id = str(uuid.uuid4())[:8]
    g.request_start_time = time.time()
    g.rate_limit = None

    rule = get_rate_limit_rule(request.method, request.path)
    if rule:
        allowed, remaining, reset_time = check_rate_limit(
            request.remote_addr,
            request.path,
            limit=rule['limit'],
            window=rule['window']
        )
        g.rate_limit = {
            'limit': rule['limit'],
            'remaining': remaining,
            'reset_time': reset_time,
        }

        if not allowed:
            response = jsonify({
                'error': 'Rate limit exceeded',
                'reset_time': reset_time,
                'retry_after': max(0, reset_time - int(time.time()))
            })
            response.status_code = 429
            response.headers['X-RateLimit-Limit'] = str(rule['limit'])
            response.headers['X-RateLimit-Remaining'] = '0'
            response.headers['X-RateLimit-Reset'] = str(reset_time)
            response.headers['Retry-After'] = str(max(0, reset_time - int(time.time())))
            return response

    if request.method in ('POST', 'PUT', 'DELETE') and request.path.startswith('/api/'):
        if request.path == '/api/csrf-token':
            return None

        csrf_token = request.headers.get('X-CSRF-Token', '')
        if not csrf_token or not verify_csrf_token(csrf_token):
            return jsonify({'error': 'Missing or invalid CSRF token'}), 403


@app.after_request
def after_request(response):
    """Post-request processing: headers, logging"""
    duration_ms = (time.time() - g.request_start_time) * 1000
    response.headers['X-Request-ID'] = g.request_id
    response.headers['X-Response-Time'] = f'{duration_ms:.2f}ms'
    response.headers['X-Service'] = 'api-gateway'

    # Security headers
    response.headers['X-Content-Type-Options'] = 'nosniff'
    response.headers['X-Frame-Options'] = 'DENY'
    response.headers['X-XSS-Protection'] = '1; mode=block'
    response.headers['Strict-Transport-Security'] = 'max-age=31536000; includeSubDomains'
    response.headers['Referrer-Policy'] = 'strict-origin-when-cross-origin'
    response.headers['Permissions-Policy'] = 'geolocation=(), microphone=(), camera=(), payment=(), usb=()'
    response.headers['Cross-Origin-Opener-Policy'] = 'same-origin'
    response.headers['Cross-Origin-Resource-Policy'] = 'same-origin'

    if getattr(g, 'rate_limit', None):
        response.headers['X-RateLimit-Limit'] = str(g.rate_limit['limit'])
        response.headers['X-RateLimit-Remaining'] = str(max(0, g.rate_limit['remaining']))
        response.headers['X-RateLimit-Reset'] = str(g.rate_limit['reset_time'])

    return response


# Health endpoints
@app.route('/health')
def health():
    """Health check"""
    return jsonify({
        'status': 'healthy',
        'service': 'api-gateway',
        'timestamp': datetime.utcnow().isoformat()
    })


@app.route('/ready')
def ready():
    """Readiness check"""
    services_status = {}
    all_healthy = True

    # Check Redis
    r = get_redis()
    if r:
        try:
            r.ping()
            services_status['redis'] = 'healthy'
        except Exception:
            services_status['redis'] = 'unhealthy'
            all_healthy = False
    else:
        services_status['redis'] = 'not_configured'

    status_code = 200 if all_healthy else 503
    return jsonify({
        'status': 'ready' if all_healthy else 'not_ready',
        'service': 'api-gateway',
        'services': services_status
    }), status_code


@app.route('/live')
def live():
    """Liveness probe"""
    return jsonify({'status': 'alive', 'service': 'api-gateway'})


@app.route('/api/csrf-token', methods=['GET'])
def csrf_token():
    """Issue a CSRF token for browser clients."""
    response = jsonify({'csrf_token': create_csrf_token()})
    response.headers['Cache-Control'] = 'no-store'
    return response


# Proxy to Auth Service (new API)
@app.route('/api/auth/<path:path>', methods=['GET', 'POST', 'PUT', 'DELETE'])
def proxy_auth(path):
    """Proxy requests to Auth Service"""
    url = f"{SERVICES['auth']}/api/auth/{path}"
    return proxy_request(url)


# Legacy Auth endpoints (for backward compatibility)
@app.route('/api/login', methods=['POST'])
def proxy_login():
    """Proxy login to Auth Service"""
    url = f"{SERVICES['auth']}/api/auth/login"
    return proxy_request(url)


@app.route('/api/register', methods=['POST'])
def proxy_register():
    """Proxy register to Auth Service"""
    url = f"{SERVICES['auth']}/api/auth/register"
    return proxy_request(url)


@app.route('/api/logout', methods=['POST'])
def proxy_logout():
    """Proxy logout to Auth Service"""
    url = f"{SERVICES['auth']}/api/auth/logout"
    return proxy_request(url)


@app.route('/api/me', methods=['GET'])
def proxy_me():
    """Proxy me to Auth Service"""
    url = f"{SERVICES['auth']}/api/auth/me"
    return proxy_request(url)


@app.route('/api/profile', methods=['PUT'])
def proxy_profile():
    """Proxy profile to Auth Service"""
    url = f"{SERVICES['auth']}/api/auth/profile"
    return proxy_request(url)


# Proxy to Product Service
@app.route('/api/products', methods=['GET'])
@app.route('/api/products/<int:product_id>', methods=['GET'])
def proxy_products(product_id=None):
    """Proxy requests to Product Service"""
    if product_id:
        url = f"{SERVICES['product']}/api/products/{product_id}"
    else:
        url = f"{SERVICES['product']}/api/products"

    return proxy_request(url)


@app.route('/api/categories', methods=['GET'])
def proxy_categories():
    """Proxy requests to Product Service"""
    url = f"{SERVICES['product']}/api/categories"
    return proxy_request(url)


@app.route('/api/homepage/content', methods=['GET'])
def proxy_homepage_content():
    """Proxy homepage content requests."""
    url = f"{SERVICES['product']}/api/homepage/content"
    return proxy_request(url)


@app.route('/api/products/<int:product_id>/reviews', methods=['GET', 'POST'])
def proxy_product_reviews(product_id):
    """Proxy product review requests."""
    url = f"{SERVICES['product']}/api/products/{product_id}/reviews"
    return proxy_request(url, require_token=(request.method == 'POST'))


@app.route('/api/wishlist', methods=['GET'])
@app.route('/api/wishlist/add/<int:product_id>', methods=['POST'])
@app.route('/api/wishlist/remove/<int:product_id>', methods=['POST'])
def proxy_wishlist(product_id=None):
    """Proxy wishlist requests."""
    if product_id and 'add' in request.path:
        url = f"{SERVICES['product']}/api/wishlist/add/{product_id}"
    elif product_id and 'remove' in request.path:
        url = f"{SERVICES['product']}/api/wishlist/remove/{product_id}"
    else:
        url = f"{SERVICES['product']}/api/wishlist"
    return proxy_request(url, require_token=True)


@app.route('/api/admin/products', methods=['GET', 'POST'])
@app.route('/api/admin/products/<int:product_id>', methods=['PUT', 'DELETE'])
@require_admin
def proxy_admin_products(product_id=None):
    """Proxy admin product management requests."""
    if product_id:
        url = f"{SERVICES['product']}/api/admin/products/{product_id}"
    else:
        url = f"{SERVICES['product']}/api/admin/products"
    return proxy_request(url, require_token=True, require_admin=True)


@app.route('/api/admin/homepage/banners', methods=['GET', 'POST'])
@app.route('/api/admin/homepage/banners/upload', methods=['POST'])
@app.route('/api/admin/homepage/banners/<int:banner_id>', methods=['PUT', 'DELETE'])
@app.route('/api/admin/homepage/offers', methods=['GET', 'POST'])
@app.route('/api/admin/homepage/offers/<int:offer_id>', methods=['PUT', 'DELETE'])
@require_admin
def proxy_admin_homepage(banner_id=None, offer_id=None):
    """Proxy homepage CMS requests."""
    path = request.path
    url = f"{SERVICES['product']}{path}"
    if path.endswith('/upload'):
        return proxy_file_upload(url, require_token=True, require_admin=True)
    return proxy_request(url, require_token=True, require_admin=True)


# Proxy to Cart Service
@app.route('/api/cart', methods=['GET', 'POST'])
@app.route('/api/cart/add/<int:product_id>', methods=['POST'])
@app.route('/api/cart/remove/<int:product_id>', methods=['POST'])
@app.route('/api/cart/clear', methods=['POST'])
@app.route('/api/cart/count', methods=['GET'])
@app.route('/api/cart/merge', methods=['POST'])
def proxy_cart(product_id=None):
    """Proxy requests to Cart Service - supports both guest and authenticated users"""
    path = request.path
    url = f"{SERVICES['cart']}{path}"
    return proxy_request(url, require_token=(path == '/api/cart/merge'))


# Proxy to Order Service
@app.route('/api/orders', methods=['GET', 'POST'])
@app.route('/api/orders/<int:order_id>', methods=['GET', 'PUT'])
@app.route('/api/orders/<int:order_id>/cancel', methods=['POST'])
def proxy_orders(order_id=None):
    """Proxy requests to Order Service"""
    if request.path.endswith('/cancel'):
        url = f"{SERVICES['order']}/api/orders/{order_id}/cancel"
    elif order_id:
        url = f"{SERVICES['order']}/api/orders/{order_id}"
    else:
        url = f"{SERVICES['order']}/api/orders"

    return proxy_request(url, require_token=True)


@app.route('/api/coupons/validate', methods=['GET'])
def proxy_coupon_validation():
    """Proxy coupon validation requests."""
    url = f"{SERVICES['order']}/api/coupons/validate"
    return proxy_request(url)


@app.route('/api/admin/coupons', methods=['GET', 'POST'])
@app.route('/api/admin/coupons/<int:coupon_id>', methods=['PUT', 'DELETE'])
@require_admin
def proxy_admin_coupons(coupon_id=None):
    """Proxy admin coupon management requests."""
    if coupon_id:
        url = f"{SERVICES['order']}/api/admin/coupons/{coupon_id}"
    else:
        url = f"{SERVICES['order']}/api/admin/coupons"
    return proxy_request(url, require_token=True, require_admin=True)


@app.route('/api/admin/orders', methods=['GET'])
@app.route('/api/admin/orders/<int:order_id>/status', methods=['PUT'])
@require_admin
def proxy_admin_orders(order_id=None):
    """Proxy admin order management requests."""
    if order_id:
        url = f"{SERVICES['order']}/api/admin/orders/{order_id}/status"
    else:
        url = f"{SERVICES['order']}/api/admin/orders"
    return proxy_request(url, require_token=True, require_admin=True)


@app.route('/api/admin/users', methods=['GET'])
@app.route('/api/admin/users/<int:user_id>/status', methods=['PUT'])
@require_admin
def proxy_admin_users(user_id=None):
    """Proxy admin user management requests."""
    if user_id:
        url = f"{SERVICES['auth']}/api/auth/admin/users/{user_id}/status"
    else:
        url = f"{SERVICES['auth']}/api/auth/admin/users"
    return proxy_request(url, require_token=True, require_admin=True)


# Proxy to Inventory Service
@app.route('/api/inventory/<int:product_id>', methods=['GET'])
def proxy_inventory_get(product_id):
    """Proxy requests to Inventory Service"""
    url = f"{SERVICES['inventory']}/api/inventory/{product_id}"
    return proxy_request(url)


@app.route('/api/admin/inventory', methods=['GET'])
@app.route('/api/admin/inventory/low-stock', methods=['GET'])
@app.route('/api/admin/inventory/restock', methods=['POST'])
@app.route('/api/admin/inventory/movements/<int:product_id>', methods=['GET'])
@app.route('/api/admin/inventory/sync', methods=['POST'])
@require_admin
def proxy_admin_inventory(product_id=None):
    """Proxy admin inventory requests."""
    path = request.path.replace('/api/admin', '/api')
    url = f"{SERVICES['inventory']}{path}"
    return proxy_request(url, require_token=True, require_admin=True)


@app.route('/api/inventory/reserve', methods=['POST'])
@app.route('/api/inventory/confirm/<int:order_id>', methods=['POST'])
@app.route('/api/inventory/release/<int:order_id>', methods=['POST'])
@app.route('/api/inventory/restock', methods=['POST'])
@app.route('/api/inventory/low-stock', methods=['GET'])
def proxy_inventory():
    """Proxy requests to Inventory Service"""
    path = request.path
    url = f"{SERVICES['inventory']}{path}"
    return proxy_request(url)


def proxy_request(url: str, require_token: bool = False, require_admin: bool = False) -> Response:
    """Proxy a request to a backend service"""
    headers = {
        'X-Request-ID': g.request_id,
        'X-Forwarded-For': request.remote_addr,
    }

    # Forward guest cart identity so cart-service can load the correct cart.
    cart_id = request.headers.get('X-Cart-ID')
    if cart_id:
        headers['X-Cart-ID'] = cart_id

    # Forward authorization header if present and attach user context when valid.
    auth_header = request.headers.get('Authorization')
    payload = None
    if auth_header:
        headers['Authorization'] = auth_header

        if auth_header.startswith('Bearer '):
            token = auth_header[7:]
            payload = verify_jwt_token(token)
            if payload:
                headers['X-User-ID'] = str(payload.get('user_id', ''))
                headers['X-User-Email'] = payload.get('email', '')
                headers['X-User-Name'] = payload.get('name', '')
                headers['X-User-Is-Admin'] = 'true' if payload.get('is_admin') else 'false'

    if require_token:
        if not auth_header or not auth_header.startswith('Bearer '):
            return jsonify({'error': 'Missing or invalid Authorization header'}), 401
        if not payload:
            return jsonify({'error': 'Invalid or expired token'}), 401
    if require_admin and not (payload and payload.get('is_admin')):
        return jsonify({'error': 'Admin access required'}), 403

    # Forward query parameters
    params = request.args.to_dict()
    json_payload = request.get_json(silent=True)
    if json_payload is not None:
        headers['Content-Type'] = 'application/json'

    try:
        if request.method == 'GET':
            resp = requests.get(url, headers=headers, params=params, timeout=30)
        elif request.method == 'POST':
            resp = requests.post(url, headers=headers, params=params, json=json_payload, timeout=30)
        elif request.method == 'PUT':
            resp = requests.put(url, headers=headers, params=params, json=json_payload, timeout=30)
        elif request.method == 'DELETE':
            resp = requests.delete(url, headers=headers, params=params, timeout=30)
        else:
            return jsonify({'error': 'Method not allowed'}), 405

        # Forward response
        response = Response(
            resp.content,
            status=resp.status_code,
            headers={'Content-Type': 'application/json'}
        )
        return response

    except requests.RequestException as e:
        return jsonify({
            'error': 'Service unavailable',
            'message': str(e),
            'request_id': g.request_id
        }), 503


def proxy_file_upload(url: str, require_token: bool = False, require_admin: bool = False) -> Response:
    """Proxy multipart form-data uploads to a backend service."""
    headers = {
        'X-Request-ID': g.request_id,
        'X-Forwarded-For': request.remote_addr,
    }

    auth_header = request.headers.get('Authorization')
    payload = None
    if auth_header:
        headers['Authorization'] = auth_header
        if auth_header.startswith('Bearer '):
            payload = verify_jwt_token(auth_header[7:])
            if payload:
                headers['X-User-ID'] = str(payload.get('user_id', ''))
                headers['X-User-Email'] = payload.get('email', '')
                headers['X-User-Name'] = payload.get('name', '')
                headers['X-User-Is-Admin'] = 'true' if payload.get('is_admin') else 'false'

    if require_token:
        if not auth_header or not auth_header.startswith('Bearer '):
            return jsonify({'error': 'Missing or invalid Authorization header'}), 401
        if not payload:
            return jsonify({'error': 'Invalid or expired token'}), 401
    if require_admin and not (payload and payload.get('is_admin')):
        return jsonify({'error': 'Admin access required'}), 403

    upload = request.files.get('image')
    if not upload:
        return jsonify({'error': 'Image file is required'}), 400

    files = {
        'image': (upload.filename, upload.stream, upload.mimetype or 'application/octet-stream')
    }

    try:
        resp = requests.post(
            url,
            headers=headers,
            files=files,
            data=request.form.to_dict(),
            timeout=60
        )
        response = Response(
            resp.content,
            status=resp.status_code,
            headers={'Content-Type': resp.headers.get('Content-Type', 'application/json')}
        )
        return response
    except requests.RequestException as e:
        return jsonify({
            'error': 'Service unavailable',
            'message': str(e),
            'request_id': g.request_id
        }), 503


if __name__ == '__main__':
    app.run(host='0.0.0.0', port=5000, debug=False)